Added All Assignments
