+++ OOP/OPP-Session-1/InterfaceAssignment/Shape.java
@@ -1,0 +1,9 @@
+/*
+ * Shape is a interface
+ * @author Banwari
+ */
+public interface Shape {
+	 
+	    public double area();
+	    public double perimeter();
+}
